refactor: move trend-search to src/lib/tools
