+++ src/lib/tools/trend-search/run.ts
@@ -1,7 +1,7 @@
-import type { TrendSearchInput, TrendSearchOutput } from "~/server/trend-search/types";
-import { planQueries } from "~/server/trend-search/query-planner";
-import { executeSearch } from "~/server/trend-search/web-search";
-import { synthesizeResults } from "~/server/trend-search/synthesizer";
+import type { TrendSearchInput, TrendSearchOutput } from "~/lib/tools/trend-search/types";
+import { planQueries } from "~/lib/tools/trend-search/query-planner";
+import { executeSearch } from "~/lib/tools/trend-search/web-search";
+import { synthesizeResults } from "~/lib/tools/trend-search/synthesizer";
 
 export type TrendSearchPipelineStage = "searching" | "synthesizing";
