+++ src/lib/tools/trend-search/synthesizer.ts
@@ -1,7 +1,7 @@
 import { ChatOpenAI } from "@langchain/openai";
 import { HumanMessage, SystemMessage } from "@langchain/core/messages";
 import { z } from "zod";
-import type { RawSearchResult, SearchCategory, SearchResult } from "~/server/trend-search/types";
+import type { RawSearchResult, SearchCategory, SearchResult } from "~/lib/tools/trend-search/types";
 
 // ─── Output schema for LLM (validate SearchResult[]) ──────────────────────────
