+++ src/lib/tools/trend-search/types.ts
