+++ src/server/inngest/functions/trendSearch.ts
@@ -1,12 +1,12 @@
 import { inngest } from "../client";
-import { runTrendSearch } from "~/server/trend-search/run";
-import { updateJobResults, updateJobStatus } from "~/server/trend-search/db";
+import { runTrendSearch } from "~/lib/tools/trend-search/run";
+import { updateJobResults, updateJobStatus } from "~/lib/tools/trend-search/db";
 import {
   TrendSearchEventDataSchema,
   type TrendSearchEventData,
   type TrendSearchOutput,
-} from "~/server/trend-search/types";
-import type { TrendSearchPipelineStage } from "~/server/trend-search/run";
+} from "~/lib/tools/trend-search/types";
+import type { TrendSearchPipelineStage } from "~/lib/tools/trend-search/run";
 
 function toErrorMessage(error: unknown): string {
   if (error instanceof Error && error.message) {
